@@ -12,6 +12,7 @@
   "files": [
     "src"
   ],
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/nrkno/valg-valgomat-algoritme.git"
@@ -31,12 +32,15 @@
     "access": "public"
   },
   "devDependencies": {
-    "eslint": "^5.16.0",
+    "@eslint/js": "^9.25.1",
+    "eslint": "^9.25.1",
     "jsverify": "0.8.4",
     "prettier": "^3.5.3",
     "tap": "^21.1.0"
   },
-  "dependencies": {},
+  "volta": {
+    "node": "22.15.0"
+  },
   "prettier": {
     "printWidth": 100,
     "semi": false,
