@@ -1,9 +1,5 @@
-function pipe(...fns) {
+export function pipe(...fns) {
   return function (x) {
     return fns.reduce((v, f) => f(v), x)
   }
-}
-
-module.exports = {
-  pipe,
 }
